#227 moved demo to packages
